// src/app/core/models/equipmentCategories.ts
export enum EquipmentCategory {
  Tractors = 'Tractors',
  Harvesters = 'Harvesters',
  PlantingEquipment = 'Planting Equipment',
  ChemicalApplicators = 'Chemical Applicators',
  TillageEquipment = 'Tillage Equipment',
  HayAndForage = 'Hay & Forage',
  Trucks = 'Trucks',
  Motorsports = 'Motorsports',
  SkidSteers = 'Skid Steers',
  Construction = 'Construction',
  OTHER = 'Other',
}